verify do not collapse on round
