+++ .gitignore
@@ -6,6 +6,7 @@ app/sign/sign
 app/stamp/stamp
 app/shamir/shamir
 app/naive/naive
+app/ntree/ntree
 deploy/platform/deterlab/build
 deploy/platform/deterlab/remote/*
 deploy/platform/deterlab/deter.toml
