+++ app/ntree/crypto.go
@@ -2,7 +2,7 @@ package main
 
 import (
 	"errors"
-	. "github.com/dedis/cothority/lib/network_draft/network"
+	net "github.com/dedis/cothority/lib/network"
 	"github.com/dedis/crypto/abstract"
 	"github.com/dedis/crypto/cipher"
 )
@@ -20,7 +20,7 @@ func hashSchnorr(suite abstract.Suite, message []byte, p abstract.Point) abstrac
 // The ring structure is removed and
 // The anonimity set is reduced to one public key = no anonimity
 func SchnorrSign(suite abstract.Suite, random cipher.Stream, message []byte,
-	privateKey abstract.Secret) BasicSignature {
+	privateKey abstract.Secret) net.BasicSignature {
 
 	// Create random secret v and public point commitment T
 	v := suite.Secret().Pick(random)
@@ -36,12 +36,12 @@ func SchnorrSign(suite abstract.Suite, random cipher.Stream, message []byte,
 	// Return verifiable si,gnature {c, r}
 	// Verifier will be able to compute v = r + x*c
 	// And check that hashElgamal for T and the message == c
-	sig := BasicSignature{Chall: c, Resp: r}
+	sig := net.BasicSignature{Chall: c, Resp: r}
 	return sig
 }
 
 func SchnorrVerify(suite abstract.Suite, message []byte,
-	signature BasicSignature) error {
+	signature net.BasicSignature) error {
 	publicKey := signature.Pub
 	r := signature.Resp
 	c := signature.Chall
