+++ deploy/monitor.go
@@ -152,7 +152,7 @@ retry:
 			k++
 			rs.StdDev = math.Sqrt(S / (k - 1))
 		} else if bytes.Contains(data, []byte("schnorr_end")) || bytes.Contains(data, []byte("end")) {
-			break
+			clientDone = true
 		} else if bytes.Contains(data, []byte("forkexec")) {
 			dbg.Lvl3("Received forkexec")
 			if rootDone {
