+++ deploy/simulation/ntree_multi_adapted.toml
@@ -1,0 +1,16 @@
+App = "ntree"
+Suite = "ed25519"
+Rounds = 10
+
+Machines, Hpn, Bf
+2, 1, 2
+4, 1, 3
+8, 1, 4
+16, 1, 5
+16, 2, 6
+16, 4, 7
+16, 8, 8
+16, 16, 9
+16, 32, 10
+16, 64, 10
+
